refactor(core): group GenomeSpy helpers
