+++ packages/core/src/genomeSpy/interactionController.js
